app/requirements.txt
python-dotenv
groq
numpy
sounddevice
webrtcvad
